Create source_to_copy.md
This .md file can be used to test and preview content changes.  This file isn't actually the source file that Github Pages used, rather you can propose changes in this file and then when we're ready to update the main landing page we copy this text (raw .md) into the actual source of the landing page, which is found under the settings tab of this repository (you have to click: settings > Launch automatic page generator to edit that one).